backend/requirements.txt
# Minimal requirements to get the API server running (Python 3.13 compatible)
# Install these first: pip install -r requirements-minimal.txt

# FastAPI & Server (updated for Python 3.13)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9  # PostgreSQL driver
pgvector>=0.2.4  # pgvector Python support

# Configuration & Validation (Python 3.13 compatible versions)
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Authentication & Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0

# AI/ML
openai>=1.12.0

# Resume/PDF Parsing
pdfplumber>=0.11.0
python-dateutil>=2.8.2

